My current code for OptimalIS
I have this optimal IS function set to print out the power set that it *would* be checking for the best IS.  It DOES NOT currently compare and return a bestIS, but making it do so would involve removing a couple comments. I do not personally see duplicates in the output, so I can not figure out why the answer *still* puts out duplicates somehow.